+++ NEWS.md
@@ -1,3 +1,25 @@
+# JWileymisc 1.2.0
+
+## Bug Fixes
+* `meanCircular()` would be off by pi in some circumstances. This has
+  been corrected and more test cases added.
+* `SEMSummary()` pairwise correlations were based on the standardized
+  pairwise covariance matrix, which used the same standard deviation
+  for a variable regardless of the pair. This has now been fixed.
+* Formula has been specified in calls to `stat_smooth()` to reduce 
+  messages about this.
+  
+## Changes
+* `corplot()` uses a more color blind friendly palette and defaults to
+  showing correlations and p-values.
+* Model diagnostics are smarter about linear regressions with essentially 
+  discrete model predictions and no longer try to run quantile regression 
+  to examine homogeneity of variance in these cases.
+* Switched to using `ggpubr` instead of `cowplot` for themes and 
+  arranging multiple plots.
+* Bumped minimum version of `R` required.
+* Further unit testing.
+
 # JWileymisc 1.1.1
 
 ## Bug Fixes